+++ app/src/androidTest/kotlin/com/k0d4black/theforce/TheForceTestApplication.kt
@@ -1,5 +1,11 @@
 package com.k0d4black.theforce
 
+import android.app.Application
+import com.k0d4black.theforce.di.fakeNetworkModule
+import com.k0d4black.theforce.di.modules.viewModelsModule
+import com.k0d4black.theforce.modules.dataSourceModule
+import com.k0d4black.theforce.modules.repositoriesModule
+import com.k0d4black.theforce.modules.useCasesModule
 import org.koin.android.ext.koin.androidContext
 import org.koin.android.ext.koin.androidLogger
 import org.koin.core.context.loadKoinModules
@@ -7,23 +13,20 @@ import org.koin.core.context.startKoin
 import org.koin.core.context.unloadKoinModules
 import org.koin.core.module.Module
 
-class TheForceTestApplication : TheForceApplication() {
+class TheForceTestApplication : Application() {
 
     override fun onCreate() {
         super.onCreate()
         startKoin {
             androidLogger()
             androidContext(this@TheForceTestApplication)
-            startKoin {
-                modules(emptyList())
-            }
+            modules(
+                fakeNetworkModule,
+                viewModelsModule,
+                repositoriesModule,
+                dataSourceModule,
+                useCasesModule
+            )
         }
     }
-
-    internal fun loadModules(module: Module, block: () -> Unit) {
-        loadKoinModules(module)
-        block()
-        unloadKoinModules(module)
-    }
-
 }
