+++ app/src/test/kotlin/com/k0d4black/theforce/activities/SearchActivityUnitTest.kt
@@ -28,7 +28,6 @@ class SearchActivityUnitTest {
                 .get()
     }
 
-
     @Test
     fun `validate default info text shown`() {
         val infoTextView = searchActivity.findViewById<TextView>(R.id.search_tip_text_view)
@@ -40,6 +39,11 @@ class SearchActivityUnitTest {
 
     @Test
     fun `validate search menu is displayed`() {
+        searchActivity =
+            Robolectric.buildActivity(SearchActivity::class.java)
+                .create()
+                .visible()
+                .get()
         val menu = Shadows.shadowOf(searchActivity).optionsMenu
         val menuItem = menu.findItem(R.id.action_search)
         assertThat(menuItem.isVisible)
